feat: categories contracts

+++ apps/back/src/services/category.ts
@@ -16,18 +16,36 @@ export const categoryService = {
     const isNotValidId = !isValidObjectId(_id);
 
     if (isNotValidId) {
-      return { category: null, isCategoryFound: false, isNotValidId };
+      return { category: null, isNotValidId };
     } else {
       const category: ICategory | null = await Category.findOne({ _id }).lean().exec();
 
-      const isCategoryFound = !!category;
-
-      return { category, isCategoryFound, isNotValidId };
+      if (category) {
+        return { category, isNotValidId };
+      } else {
+        return { category: null, isNotValidId: true };
+      }
     }
   },
 
   update: async (_id: string, categoryToUpdate: ICategory) => {
-    await Category.findOneAndUpdate({ _id }, { ...categoryToUpdate, dateUpdated: new Date() });
+    const isNotValidId = !isValidObjectId(_id);
+
+    if (isNotValidId) {
+      return { isNotValidId };
+    } else {
+      const category = await Category.findOne({ _id });
+
+      if (category) {
+        await Category.updateOne({ ...categoryToUpdate, dateUpdated: new Date() });
+
+        category.save();
+
+        return { isNotValidId };
+      } else {
+        return { isNotValidId: true };
+      }
+    }
   },
 
   create: async (categoryToCreate: ICategory) => {
@@ -47,10 +65,22 @@ export const categoryService = {
   },
 
   delete: async (_id: string) => {
-    const category = await Category.findOne({ _id });
+    const isNotValidId = !isValidObjectId(_id);
 
-    deleteFile(category?.iconUrl);
+    if (isNotValidId) {
+      return { isNotValidId };
+    } else {
+      const category = await Category.findOne({ _id });
 
-    await category?.deleteOne();
+      if (category) {
+        deleteFile(category?.iconUrl);
+
+        await category?.deleteOne();
+
+        return { isNotValidId };
+      } else {
+        return { isNotValidId: true };
+      }
+    }
   },
 };
